Change to db schema. Everything works. Also fresh sqldump
