+++ SQL/realestate_structure_cleaned.sql
@@ -1,16 +1,27 @@
+-- MySQL dump 10.13  Distrib 8.0.19, for osx10.15 (x86_64)
+--
+-- Host: localhost    Database: realestate
+-- ------------------------------------------------------
+-- Server version	8.0.19
 
-use realestate;
+/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
+/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
+/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
+/*!50503 SET NAMES utf8mb4 */;
+/*!40103 SET @OLD_TIME_ZONE=@@TIME_ZONE */;
+/*!40103 SET TIME_ZONE='+00:00' */;
+/*!40014 SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0 */;
+/*!40014 SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0 */;
+/*!40101 SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='NO_AUTO_VALUE_ON_ZERO' */;
+/*!40111 SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0 */;
 
--- execute SET/DROPS one at a time
-SET FOREIGN_KEY_CHECKS = 0;
+--
+-- Table structure for table `address`
+--
+
 DROP TABLE IF EXISTS `address`;
-DROP TABLE IF EXISTS `house_photos`;
-DROP TABLE IF EXISTS `listing_detail`;
-DROP TABLE IF EXISTS `realtor`;
-SET FOREIGN_KEY_CHECKS = 1;
-
-
--- execute CREATE TABLE one at a time
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `address` (
   `house_number` int DEFAULT NULL,
   `street` varchar(100) COLLATE utf8mb4_general_ci DEFAULT NULL,
@@ -20,18 +31,16 @@ CREATE TABLE `address` (
   `zip` int NOT NULL,
   `apt_suite` varchar(100) COLLATE utf8mb4_general_ci DEFAULT NULL,
   PRIMARY KEY (`id`)
-) ENGINE=InnoDB AUTO_INCREMENT=5 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
+) ENGINE=InnoDB AUTO_INCREMENT=13 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
 
+--
+-- Table structure for table `listing_detail`
+--
 
-CREATE TABLE `realtor` (
-  `id` int NOT NULL AUTO_INCREMENT,
-  `name` varchar(100) COLLATE utf8mb4_general_ci NOT NULL,
-  `description` varchar(500) COLLATE utf8mb4_general_ci NOT NULL,
-  `picture` mediumblob,
-  PRIMARY KEY (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
-
-
+DROP TABLE IF EXISTS `listing_detail`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `listing_detail` (
   `id` int NOT NULL AUTO_INCREMENT,
   `description` varchar(1000) COLLATE utf8mb4_general_ci DEFAULT NULL,
@@ -40,20 +49,42 @@ CREATE TABLE `listing_detail` (
   `realtor_id` int NOT NULL,
   `price` int DEFAULT NULL,
   `adress_id` int NOT NULL,
+  `photo` mediumblob,
   PRIMARY KEY (`id`),
   KEY `listing_detail_FK_1` (`adress_id`),
   KEY `listing_detail_FK` (`realtor_id`),
   CONSTRAINT `listing_detail_FK` FOREIGN KEY (`realtor_id`) REFERENCES `realtor` (`id`),
   CONSTRAINT `listing_detail_FK_1` FOREIGN KEY (`adress_id`) REFERENCES `address` (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
+) ENGINE=InnoDB AUTO_INCREMENT=9 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
 
+--
+-- Table structure for table `realtor`
+--
 
-CREATE TABLE `house_photos` (
+DROP TABLE IF EXISTS `realtor`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!50503 SET character_set_client = utf8mb4 */;
+CREATE TABLE `realtor` (
   `id` int NOT NULL AUTO_INCREMENT,
-  `listing_id` int DEFAULT NULL,
-  `is_default` tinyint(1) DEFAULT NULL,
-  `photo` mediumblob,
-  PRIMARY KEY (`id`),
-  KEY `housephotos_FK` (`listing_id`),
-  CONSTRAINT `housephotos_FK` FOREIGN KEY (`listing_id`) REFERENCES `listing_detail` (`id`) ON DELETE CASCADE ON UPDATE CASCADE
-) ENGINE=InnoDB AUTO_INCREMENT=3 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
+  `name` varchar(100) COLLATE utf8mb4_general_ci NOT NULL,
+  `description` varchar(1000) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci DEFAULT NULL,
+  `picture` mediumblob,
+  PRIMARY KEY (`id`)
+) ENGINE=InnoDB AUTO_INCREMENT=12 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Dumping routines for database 'realestate'
+--
+/*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
+
+/*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
+/*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
+/*!40014 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS */;
+/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
+/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
+/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
+/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
+
+-- Dump completed on 2020-02-29 12:07:57
