+++ WebContent/listing-form.jsp
@@ -89,5 +89,25 @@
 				name="description"></textarea>
 		</div>
 	</div>
+	<!-- File Button -->
+
+	<label
+		class="label"
+		for="photo">Photo</label>
+	<div class="file">
+		<label class="file-label"> <input
+			class="file-input"
+			type="file"
+			name="photo"
+			onchange="if (this.files.length > 0) document.getElementById('filename-photo').innerHTML = this.files[0].name;">
+			<span class="file-cta"> <span class="file-icon"> <i
+					class="fa fa-upload"></i>
+			</span> <span
+				class="file-label"
+				id="filename-photo"> Choose a file... </span>
+		</span>
+		</label>
+	</div>
+
 
 </fieldset>
