+++ src/main/java/controller/ListingDetailDAO.java
@@ -11,7 +11,6 @@ import javax.persistence.Persistence;
 import javax.persistence.TypedQuery;
 
 import model.Address;
-import model.HousePhoto;
 import model.ListingDetail;
 
 /**
